wr_clocksource: Fix bug due to clock adjustment
The WR time adjustment was not taken into account in the driver. The
side effect was a big jump sometime of the monotonic clock used in PPSI
for the calculation of the timeouts. The side effect was that many
timeouts were firing too early.
The driver registration take now into account the adjustment and also
the calculation of the offset in the read() call.