[issue #20] output syslog to /tmp with rotation of files
rsyslog outputs syslog to /tmp/syslog until the file is ~1MB, then
the file is moved/renamed to a "rotate file" (name includes date &
time) while the /tmp/syslog is filled by rsyslog from zero. Max
number of 10 "rotate files" are kept. The oldest is removed when
new file from rsyslog comes. Next step is to make this a default
option in kconfig