[issue #194] update init scripts to provide name of deamon in syslog
Init scripts start logger that prints output of deamons into syslogs.
This logger is run with -p to attach tag. This tag is interpreted
as the process that produces the following message. In the many init
scripts this tag was set to wr-switch, incorrectly. This commit
corrects it to include the name of the daemon.