[issue #194] add facility to wrs_msg and <facility*8+priority> to output msgs
Logger can interpret get log facility & level (error, info, etc) from
log message if such info is provided properly. I.e. it needs to be
enclosed in <> and the first thing in the message. The wrs-msg.c file
in the libwr was changed such that the log facility & level ares properly
printed in log messages, and thus later properly interpreted by the syslog
elastic search on the server. For this, wrs_msg lib needs to know the
faciltiy of the application that uses it. Thus, the wrs_msg_init()
function was updated (in the library and all its calls in applications)