wrs_build_all: Exit on error
Having just a small message indicating that a build error has been
detected at the end of the script is not sufficient. The tarball is
created but is invalid and might be used.
Now, when an error is detected the build procedure is immediately
aborted. The tarball is not then created.