wr_date.c: Add option
The -n option in the wr_date tool is used for a dry run. It execute the
code without applying changes in the system and the hardware.
However calling "wr_date -n set host" was not setting the date but
applying the TAI offset. To make things more consistant, a new option
has been added. Calling "wr_date set host tai" now sets only the TAI
offset.
The wr_date script has been changed to take into account this
modification of the behavior.