[issue #194] add optiont to Kconfig which allows using <facility*8+priority> from messages
Another option is added in Kconfig that applies to all the deamons
that provide <facility*8+priority> in their messages. If logging
in Kconfig is set to "default_syslog", the logger is run with
the option --prio-prefix. With this option, the facility/priority
from message is used. All the other otions for "Logint directions.."
are left unchnaged.