[issue #8] Avoid locking already locked shmem
Shmem is locked for the entire execution of the rtu_fd_init()
function. Yet, this function call few clean_*() functions that
attempt to lock shmem before cleaning... this caused shmem
errors. It was fixed by specifying input parameter to clean_*()
functions. The parameter specifies whether locking is required
or not.