[issue #11] Prevent static MAC entries from removing by dynamic entires
If a static entry is in the HASH table, this is for a reason. In the
very unlikely event, when the same MAC is seen on another port,
the static entry should not be overriden by the learned (dynamic).
This was not the case. This commit prevents static MAC entries
from being overriden by dynamic entries.