SNMP: fixed/updated wrsPtpDataTable
The fields in this table were not really filled in after
substantial changes to the PPSi architecture regarding
servo (HA-related). Filling in of the fields in this
table had to be redone. This commit fixes filling in
most of the fields that were field before. There are still
some fields that are not filled in, tbd.