[HAL] correct behavior for SFPs with corrupted header.
Despite corrupted header, we still want to use this
SFP because it can actually be functional. We have
copper SFPs with such problem that are fully functional
(and did work with the previous version of HAL, by chance).

So, we now detect insertion of such SFP (with unreadable
header), yet we make sure that such SFP is declared
unrecognized (zero header data) such that SNMP complains.