[HAL] read LPCD tx calibration file only once, do it when initializing all ports
If already exists, the file needs to be read once for all ports and
the structures of the ports need to be updated. this is better done
at the very begining for all ports rather then making a hack and
executing it when port_index=0. Thus, move the thing to the initial
global init.