[HAL] update of _pll_state by each port.
As it is now, once the port with index=0 is calibrated, the
_pll_state is never updated. Also, since the states are called
recursively in case new state is entered, the _pll_state is not
updated unless this is called on port with index=0.

Trying to call it regardless of the port before implementing
complicated mechanism to manage updates once per all ports under
calibration.