[HAL] bugfix: prevent reseting TIMING_MODE if non-slave port is unplugged
If there were more ports connected than just slave port, unplugging any
of the ports would result (on BC) the timing_mode to be changed to
FREE_MASTER. As a consquence, the port in slave state was reseted and
would re-establishs sync. It was fixed by checking whether the unplugged
port was slave (i.e. was locked)