[HAL] add timer in rx_setup FSM to wait for early_link_up to be cleared.
After unplugging fiber, the early_link_up flag is cleared a bit later
than the detection of link_down happens. This resulted in the tx_setup
state machine trying to calibrate unplugged link. The things worked
because the FSM would hang in wait_lock state and then timeout. Yet,
this did not seem the best solution. So, I added timeout to wait
for the early_link_up bit to be cleared