dot-config script: new behavior
When the dot-config file is read at startup using the DHCP service to
get the server IP address, the script waits forever the availability of
DHCP service in the following case :
- CONFIG_DOTCONF_SOURCE_FORCE_DHCP=y AND CONFIG_ETH0_DHCP=y

CONFIG_ETH0_DHCP: Set to indicate that the network is forced to use DHCP
for its setup.

In the other cases, the behavior stay unchanged.