Kconfig: new dependency
A dependency have been added to not allow the option "Force to get the
URL to a dot-config via DHCP" in order to get the source of the
dot-config file. This option is now only available if the "DHCP forever"
option is selected in the management port configuration.

This has been done to avoid to wait forever the availability of the DHCP
service to get the dot-config source while this service is not required
or mandatory for the management port setup.