[snmp] Bugfix: do not check startCnt when daemon is disabled.
SNMP checkes whether monitored daemons where restarted by looking
at the respective /tmp/start_cnt_X file. Such file is not created
if a daemon is disabled in configuration (dotconfig). Such
disabling is a new feature. It turns out that not all of SNMP
deamon was updated to accommodate it. The SNMP daemon was checking
at start which daemons are disbled but later did not use this
info in wrsStartCntGroup and the startCnt of the disabled deamons
was attempted to be checked, resulting in an error. This has
been fixed.