[syslog] specify default facility.priority in case it is not included in message
Some deamons/etc that log messages to syslog do not specify facility.level
in their messages. The logger is configured to parse the message looking
for facility.level (option: --prio-prefix). If the message does not include
If no facility.priority is specified, the logger will use the one specified
after -p, this is now provided. For example, PPSi does not specify
facility.priority in messages (hopefully, this will change), without
"-p daemon.info", it was reported as user.info. If user wishes to,
he/she can override. This is the default config