[HAL] turn calibration LED off when starting calibration
In the rx calib FSM, this is needed to over the case when the
calibration is re-started (INIT state entered) during execution
of the FSM (e.g. due to link unplugging, or timeout).

In the tx calib FSM, in theory this is not needed, yet it seems
clean and it might be needed in the future if some sort of
reset is implemented.