Fix monit problem with commit 9bc9dd1
There is a monit-related problem in commit 9bc9dd1.

The commit renames /etc/init.d/lldpd.sh into /etc/init.d/lldpd.bash
whle adding a single stanza. Try "git diff e61eb08 9ebc5e2" to
see the difference.

But monit still uses the old name. Instead of changing monit
(a two-liner, but adds noise because of .bash), this renames back
the file to .sh and removes the needless bashism in the new stanza.

Please note that calling stuff ".bash" makes no sense (suffixes
make no sense most of the time!). The only thing rcS does is checking
if the name is "*.sh", to just read it in the current shell instead
of executing. So anything not ".sh" is executed, we don't need to
special case .bash, .py or whatever.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>