snmp/bridge_mib: fix length of dot1dStaticTable as an index
In the generated code MacAddress is mapped into ASN_OCTET_STR, which uses
a length as a first byte, then the content of MAC/string.
With this behaviour of ASN_OCTET_STR, the index was 1 byte longer than suppose
to be. Anyway, it is not needed to carry the length of MacAddress since it is
defined as fixed in the MIB.

Signed-off-by: Michal Wasiak <michal.wasiak@gmail.com>