/etc/init.d/lldpd.sh: fix another bashism
[Thanks a lot to Enkhbold Ochirsuren of GSI who traced back a strange
behaviour to the "printf" incompatibility].

This removes "printf -v <outputvar> <format> ..." which is only
available in bash.

These needless incompatible extensions to standard commands are a
source of endless headaches. New features must spit an error when
unavailable (i.e., bash people should add vprintf, not break printf,
imho)

In this case, my previous "fix" to this file was incomplete, because
I fixed a bashism but didn't see this other one.

Also, this changes the shebang to call sh (which is already done by rcS:
only the interactive execution now uses sh instead of bash).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>