[DOC #362] doc/wrs-user-manual: reorganizse the documentation.
The part of the document that describes each configuration item
(chapter "Configuration of the Device") is separated from the
part that describes more in details behavior of the device,
give more info about the configuration itself, the relations
between items, etc. It is now in the chapter "Operation of the
Device".

NOTE: I tried to take the content of Apendix B on time and
frequency sources and inject it in this chapter under "Time
and frequency sources of Grandmaster" but the result required
too many levels of embedded sub-sections (not supported) and
IMHO it is important that one can find reference to this info
direclty in the index page.